refactor(agent-host): split the execution coordinator and reducer off the filesize allowlist
projectAgentExecutionCoordinator.ts 1839 -> 774 and projectAgentReducer.ts 936 -> 784. Both
leave the allowlist. Authored by a Codex worker in an isolated worktree; committed here because
its sandbox cannot write the linked worktree's git index.

Seams, all chosen off the per-record path:
- projectAgentExecutionCoordinatorTypes.ts (281): shared coordinator contract/state ownership.
- projectAgentExecutionRecovery.ts (135): process-restart and orphan recovery during partition
  initialization — a lifecycle concern, runs once.
- projectAgentAdapterResolvers.ts (122): per-turn/tool transport selection, a setup boundary.
- projectAgentTurnExecution.ts (745): the per-turn execution flow.
- projectAgentThreadReduction.ts (221): thread lifecycle plus turn.enqueue — thread setup and
  queue admission, not per-record assertion work.

Why the seam choice was the whole job here, measured rather than assumed:
This pair has already produced both a regression and a win this session, and neither was
visible in typecheck or in pass/fail. Splitting projectAgentReducer.ts ALONE onto an older main
took the reducer perf assertion from 18847ms to 32329ms against an 8000ms budget — 1.7x worse,
with tests still "passing" in the sense that the same set failed. A sibling file,
projectAgentState.ts, was split by lifting its per-record assertion layer and a hot-path test
went 25s -> 47s and timed out; that one is still allowlisted and should stay that way. The rule
that came out of it: seams on a per-record loop cost real time, seams at per-turn / setup /
transport boundaries do not.

Measured before and after with the identical command, npx vitest run electron/projectAgentHost/,
and re-measured independently rather than taking the worker's numbers:
- pristine main: 261 passed / 1 skipped, Duration 21.14s
- with both splits: 261 passed / 1 skipped, Duration 19.99s
Same test counts, same (empty) failure set, no timeouts, no assertion weakened, no timeout
raised. The directory is fully green on current main — note this is a different baseline from
earlier today, because main has since reworked those heavy tests, which is exactly why each
seam has to be re-measured rather than trusted from a previous branch.

Verified: pnpm run typecheck EXIT=0; pnpm run lint:ci EXIT=0 (82 warnings, at the threshold —
the imports the extraction left behind were removed); node scripts/check-file-sizes.mjs EXIT=0;
electron/projectAgentHost/ 261 passed. Line counts re-measured independently.
